Extract WireData to sharedTypes file
